refactor game objects to setup
